Enable configuration of mathfu dependencies.
This commit allows users of cmake and cmake GUI to configure the location of
the library's dependencies using the dependencies_* group of variables.
When displaying the variables in cmake GUI they're shown as a group under
"dependencies" which allows the user to select directories for the location
of the googletest and vectorial projects.

Tested:
* Verified all projects still build successfully
* Verified the dependencies_* variables are displayed in cmake GUI.

Bug: 16630244
Change-Id: I42f86aef94590b89644ccd3d6a7be29913788a17