Added SIMD and non-SIMD build targets for benchmarks and tests.
When SIMD is enabled for MathFu, both SIMD and non-SIMD binaries are built for
benchmarks and unit tests.  This allows comparison between the performance of
SIMD vs. non-SIMD builds of benchmarks.  In addition, this makes it possible
to test the non-SIMD code paths on platforms where SIMD is enabled by default
(Windows, OSX, Linux).

Tested:
Verified all targets build on OSX, Windows and Linux.
Inspected generated binaries and verified SIMD instructions are used in the
SIMD builds vs. non-SIMD builds.

Bug: 16626436
Change-Id: I4e0f8ad9b8c096504fac60f6dbd05810e3a15ee5