Added options to disable build of benchmarks and tests.
Tests were always built when the cmake project is included.
This isn't great for users of the project who probably do not want
to build all the tests when including MathFu in their project.

Bug: 17456339

Tested:
Verified that benchmarks and tests no longer build when the options
are turned off on Linux.

Change-Id: I5d3cd9b29b74daa0a2d3f116fcc99db611434205