Remove all instances of MATHFU_CAST.
MATHFU_CAST is causing problems for older compilers like gcc 4.6
on Linux and with the Android toolchain.

In addition, simplified some duplicated code and modified the
matrix benchmark so that it's possible to change the number
of dimensions of the matrix.

Tested:
Verified all tests build and pass on Android
(for all ABIs - only tested ARM), Linux, OSX and Windows.

Bug: 17370290
Change-Id: I31d4856702718245afac7fc0cb2f8f71e8c28830