Implemented packed vector structure and serialization routines.
It is sometimes useful to represent vectors as simply arrays of
floats in order to send them to a GPU.  The library pads vector
classes in some build configurations so that it's possible to use
SIMD operations to increase performance.  This change introduces
a packed vector structure and methods to convert from unpacked
vector classes or matrices to this structure.

Tested:
Verified all tests build and pass on OSX.

Bug: 17529907
Change-Id: If2fa32ceeb8089bc9bf82f5998f6ff1ff82d3bf3