Cleaned up and expanded random functions
Added a function that lets you specify a range of numbers, a-b,
and returns a random number in the range [a, b)

Also added a vector version which, given two vectors, returns a
third vector whose elements are bounded by the corresponding
elements of the first two.

Also modified the base Random() function to return numbers that
are >= 0 and < 1.  (Before they could sometimes equal 1)

Change-Id: Ibda71acc9a7a12befa97e799b377958e9abc5502