Added quaternion accessors and mutators.
Sometimes it's convenient to have direct access to the scalar and vector
components, so I added accessors and mutators for both, along with unit
tests for each accessor and mutator.

This makes it easier to do things like this:

    Vector<float, 3> my_vec = quaternion.vector();

Rather than needing to do this:

    Vector<float, 3> my_vec(quaternion[1],
                            quaternion[2],
                            quaternion[3]);

Tested on Linux, all unit tests pass.

Change-Id: Ic4ae021dcc6439deb6d80d35a444a24afa343533