Fix Matrix<>::operator*().
The underlying Vectorial call needs different references for
input and output matrices, but we were giving it 'this.data_'
for both.

Fixed by making a copy of *this and passing that in as the
input parameter. In practice, this copy shouldn't cost anything.
The stores to the stack should be optimized away.

The non-SIMD version was broken in the same way, and fixed
in the same way.

Bug: 19246559
Change-Id: I99cd7ec4206334fb799416f96db82893ecb5088b
Tested: Added tests for Matrix multiplications.