Fix directory reference.
This was working only by co-incidence, since the mathfu and
external directories happened to be at the same level.

This path is relative to mathfu's CMakeList.txt directory,
however, so we need to use CMAKE_CURRENT_SOURCE_DIR
instead of CMAKE_SOURCE_DIR.

Tested: Works on mathfu, motive, and pie_noon on Linux.
Change-Id: I80afb89d32bbff9aa98ffd51c940b8b29d0762c4