Adding AffineTransform functions to Mathfu.
In order to facilitate the changes to shader uniforms now using
Matrix<float, 4, 3>, the functions to convert to/from this type
needed to be in a central location. Mathfu is a logical choice,
since the conversions are between mathfu types, and the library is
common amongst all affected libraries.

Bug: b/24951383
Change-Id: I6f42504bc45bef7128e12087d112655ecdee9196
Tested: Tested by building and running Zooshi on Ubuntu.