Find dependencies directory when mathfu already in it.
This fixes motive/src/anim_pipeline/CMakeLists.txt's build
with the `dependencies` directory layout. The problem was
that the dependencies directory is,
  motive/dependencies
but mathfu looks in
  motive/src/anim_pipeline/dependencies
which doesn't exist.

When the dependencies directory doesn't exist, we can't
find the vectorial directory, and builds fail.

Tested: Linux builds of anim_pipeline work with `dependencies`
directory. Linux builds of mathfu are unaffected.

Change-Id: Ie444f416548bf0816f4cb9f0e6154f5b40a405d2