Add handedness to the LookAt() and Ortho() functions.
Note--this change does not affect functionality at all.
All results should remain the same.

Unfortunately, our pre-existing LookAt() is left-handed.
This is different from the defaults for Perspective() and Ortho(),
which are right-handed.

For now, the default for LookAt() remains left-handed, so that
this change should break no code.

However, we should, at some point, introduce a breaking change
that makes the default right-handed so that our functions match.

Also, in PerspectiveHelper(), remove double multiply of `handedness`
in the last term. That term is independent of `handedness`.
Note--the results are the same since we were multiplying the last
term by handedness*handedness = 1 (we've removed this).

Added unit tests
Bug: 19229170

Change-Id: I1df7c591ed42dbee6071db8fc27d67db3b6255ee