Add FromType() and ToType() from/to byte-wise compatible conversions.
Performs safe and efficient casts between Matrix and Vector
types with non-mathfu Matrix and Vector types.

Useful because there are so many vector types in the world,
and reinterpret_cast is so tempting, and so subtly wrong do to
aliasing bugs.

Change-Id: Ia24f432bd6fe97df434fa0b321da65b050572999
Tested: New unit tests pass on Linux, Mac, and Windows.