Include template specializations in template headers.
Previously, the code below wouldn't work,

  #include "vector.h"
  mathfu::Vector<float, 3> myVec;
  #include "vector_3.h"

It would generate compile error,
"specialization of ‘mathfu::Vector<float, 3>’ after instantiation
class Vector<float, 3>"

The solution is to include vector_3.h in vector.h, and similarly
for the other specializations.

Note that, if you don't want specializations, you can disable them
by defining MATHFU_COMPILE_WITHOUT_SIMD_SUPPORT.

We hide the specializations in an 'internal' directory because
there are include order dependencies. By allowing the user to
explicitly include one of the specializations, the include order
can be messed-up. For example,
  #include "vector_3.h"
would have included vector.h, which in-turn includes vector_3.h
again, but vector_3.h is not processed because we're already inside
of it. Then vector.h includes vector_4.h, which fails because it
has a dependency on vector_3.h (in the vec3,float constructor).

The fix is to have vector_3.h redirect to vector.h, and move the
specializations to internal header files so that they're never
explicitly included.

Change-Id: Ie9ddaa4c89e9eb1e43828c5100f8626c97e7bbeb
Tested: Unit tests build and pass on Linux.