Suppress excessive warnings in cmake and Windows builds.
Some builds can use GCC instead of clang, so we need to suppress
the "anonymous struct/unions are not ANSI-C" warning for GCC as well.

Windows had an "array index out of bounds" warning that we were
suppressing in vector.h, but not in internal/vector_2.h, etc.
Ensure that suppression continues.
PiperOrigin-RevId: 141503071
Change-Id: I815070d64cb657306038cf6054d14268d59be5d4