Loosen epsilon in Quaternion::Slerp.
If very similar q1 and q2 are passed in, they can fail the "> 0.999999f" check and yet still result in NaNs returned from the expression:
((q1.Inverse() * q2) * s1)
on the last line.
(q1.Inverse() * q2) is a valid quaternion, but multiplying it by s1 results in NaNs (regardless of the value of s1).

PiperOrigin-RevId: 215971592
Change-Id: I1dcdf713c903213cef42f52dfe89fe71e5231293