Add Quaternion::operator*(float) to the Slerp unit test
Requires fixing a NaN edge case in operator*.  The root cause was
ToAngleAxis not checking the result of Vector::Normalize().

For stability, the zero-length check in ToAngleAxis could be improved
into an epsilon check; in unstable cases the resulting axis may be
slightly non-unit.  I think it's not a practical issue so didn't
bother addressing it.

The current implementation of quat*float handles the instability by
re-normalizing the axis that comes back from ToAngleAxis().  This
works fine, but is a little wasteful in general.

PiperOrigin-RevId: 217628403
Change-Id: Ibd2079dd266934daada2e3ed615c973fbc340832